:twisted_rightwards_arrows: Split compilation into sub classes
