+++ packages/pundle/src/helpers.js
@@ -3,12 +3,10 @@
 /* @flow */
 
 import Path from 'path'
-import sourceMap from 'source-map'
-import isRegexp from 'lodash.isregexp'
 import FileSystem from 'pundle-fs'
 import { parse } from 'babylon'
 import type Pundle$Path from './path'
-import type { Pundle$Config, Pundle$Plugin, Pundle$FileSystem, Pundle$Watcher$Options } from './types'
+import type { Pundle$Config, Pundle$Plugin, Pundle$FileSystem } from './types'
 
 const REGEX_EOL = /\n|\r\n/
 
@@ -48,18 +46,6 @@ export function normalizeConfig(givenConfig: Pundle$Config): Pundle$Config {
   }
   config.sourceMaps = Boolean(config.sourceMaps)
   return config
-}
-
-export function normalizeWatcherOptions(givenOptions: Object): Pundle$Watcher$Options {
-  const options = Object.assign({}, givenOptions)
-  if (typeof options.ignored !== 'string' && !isRegexp(options.ignored) && !Array.isArray(options.ignored)) {
-    options.ignored = /(node_modules|bower_components)/
-  }
-  if (typeof options.onError !== 'function') {
-    throw new Error('options.onError must be a function')
-  }
-  options.ignored = [/[\/\\]\./].concat(options.ignored)
-  return options
 }
 
 export async function find(
@@ -122,45 +108,6 @@ export async function getPlugins(
     processed.push({ plugin, parameters })
   }
   return processed
-}
-
-// Source: https://goo.gl/821D9T
-export function mergeSourceMaps(inputMap: Object, map: Object): Object {
-  const inputMapConsumer   = new sourceMap.SourceMapConsumer(inputMap)
-  const outputMapConsumer  = new sourceMap.SourceMapConsumer(map)
-
-  const mergedGenerator = new sourceMap.SourceMapGenerator({
-    file: inputMapConsumer.file,
-    sourceRoot: inputMapConsumer.sourceRoot
-  })
-
-  // This assumes the output map always has a single source, since Babel always compiles a single source file to a
-  // single output file.
-  const source = outputMapConsumer.sources[0]
-
-  inputMapConsumer.eachMapping(function(mapping) {
-    const generatedPosition = outputMapConsumer.generatedPositionFor({
-      line: mapping.generatedLine,
-      column: mapping.generatedColumn,
-      source
-    })
-    if (typeof generatedPosition.column !== 'undefined') {
-      mergedGenerator.addMapping({
-        source: mapping.source,
-
-        original: {
-          line: mapping.originalLine,
-          column: mapping.originalColumn
-        },
-
-        generated: generatedPosition
-      })
-    }
-  })
-
-  const mergedMap = mergedGenerator.toJSON()
-  inputMap.mappings = mergedMap.mappings
-  return inputMap
 }
 
 export function getLinesCount(text: string): number {
