+++ packages/pundle/src/processor/helpers.js
@@ -2,7 +2,47 @@
 
 /* @flow */
 
+import sourceMap from 'source-map'
 import { VISITOR_KEYS } from 'babel-types'
+
+// Source: https://goo.gl/821D9T
+export function mergeSourceMaps(inputMap: Object, map: Object): Object {
+  const inputMapConsumer   = new sourceMap.SourceMapConsumer(inputMap)
+  const outputMapConsumer  = new sourceMap.SourceMapConsumer(map)
+
+  const mergedGenerator = new sourceMap.SourceMapGenerator({
+    file: inputMapConsumer.file,
+    sourceRoot: inputMapConsumer.sourceRoot
+  })
+
+  // This assumes the output map always has a single source, since Babel always compiles a single source file to a
+  // single output file.
+  const source = outputMapConsumer.sources[0]
+
+  inputMapConsumer.eachMapping(function(mapping) {
+    const generatedPosition = outputMapConsumer.generatedPositionFor({
+      line: mapping.generatedLine,
+      column: mapping.generatedColumn,
+      source
+    })
+    if (typeof generatedPosition.column !== 'undefined') {
+      mergedGenerator.addMapping({
+        source: mapping.source,
+
+        original: {
+          line: mapping.originalLine,
+          column: mapping.originalColumn
+        },
+
+        generated: generatedPosition
+      })
+    }
+  })
+
+  const mergedMap = mergedGenerator.toJSON()
+  inputMap.mappings = mergedMap.mappings
+  return inputMap
+}
 
 export function getName(obj: Object): string {
   if (typeof obj.name === 'string') {
